Make CreateDagVisitor more generic
