+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/opt/physical/AggregateByKeyPhysicalRel.java
@@ -18,7 +18,6 @@ package com.hazelcast.jet.sql.impl.opt.physical;
 
 import com.hazelcast.function.FunctionEx;
 import com.hazelcast.jet.aggregate.AggregateOperation;
-import com.hazelcast.jet.core.Vertex;
 import com.hazelcast.jet.sql.impl.ObjectArrayKey;
 import com.hazelcast.jet.sql.impl.opt.OptUtils;
 import com.hazelcast.sql.impl.QueryParameterMetadata;
@@ -66,7 +65,7 @@ public class AggregateByKeyPhysicalRel extends Aggregate implements PhysicalRel 
     }
 
     @Override
-    public Vertex accept(CreateDagVisitor visitor) {
+    public <V> V accept(CreateDagVisitor<V> visitor) {
         return visitor.onAggregateByKey(this);
     }
