+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/opt/physical/AggregateCombineByKeyPhysicalRel.java
@@ -17,7 +17,6 @@
 package com.hazelcast.jet.sql.impl.opt.physical;
 
 import com.hazelcast.jet.aggregate.AggregateOperation;
-import com.hazelcast.jet.core.Vertex;
 import com.hazelcast.jet.sql.impl.opt.OptUtils;
 import com.hazelcast.sql.impl.QueryParameterMetadata;
 import com.hazelcast.sql.impl.plan.node.PlanNodeSchema;
@@ -60,7 +59,7 @@ public class AggregateCombineByKeyPhysicalRel extends Aggregate implements Physi
     }
 
     @Override
-    public Vertex accept(CreateDagVisitor visitor) {
+    public <V> V accept(CreateDagVisitor<V> visitor) {
         return visitor.onCombineByKey(this);
     }
