+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/opt/physical/PhysicalRel.java
@@ -16,7 +16,6 @@
 
 package com.hazelcast.jet.sql.impl.opt.physical;
 
-import com.hazelcast.jet.core.Vertex;
 import com.hazelcast.jet.impl.util.Util;
 import com.hazelcast.jet.sql.impl.opt.OptUtils;
 import com.hazelcast.jet.sql.impl.schema.HazelcastTable;
@@ -71,5 +70,5 @@ public interface PhysicalRel extends RelNode {
      * @param visitor Visitor.
      * @return the DAG vertex created for this rel
      */
-    Vertex accept(CreateDagVisitor visitor);
+    <V> V accept(CreateDagVisitor<V> visitor);
 }
