+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/opt/physical/ShouldNotExecuteRel.java
@@ -17,7 +17,6 @@
 package com.hazelcast.jet.sql.impl.opt.physical;
 
 import com.hazelcast.internal.util.StringUtil;
-import com.hazelcast.jet.core.Vertex;
 import com.hazelcast.jet.sql.impl.opt.OptUtils;
 import com.hazelcast.sql.impl.QueryException;
 import com.hazelcast.sql.impl.QueryParameterMetadata;
@@ -67,7 +66,7 @@ public class ShouldNotExecuteRel extends AbstractRelNode implements PhysicalRel 
     }
 
     @Override
-    public Vertex accept(CreateDagVisitor visitor) {
+    public <V> V accept(CreateDagVisitor<V> visitor) {
         throw QueryException.error(message());
     }
