+++ hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/opt/physical/CreateDagVisitorTest.java
@@ -38,7 +38,7 @@ public class CreateDagVisitorTest {
         Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
         Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
         dag.edge(between(a, b));
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(a.getLocalParallelism(), 4);
         assertEquals(b.getLocalParallelism(), 4);
@@ -52,7 +52,7 @@ public class CreateDagVisitorTest {
         Vertex c = dag.newVertex("c", PROCESSOR_SUPPLIER);
         dag.edge(between(a, b));
         dag.edge(between(b, c));
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(a.getLocalParallelism(), 4);
         assertEquals(b.getLocalParallelism(), 4);
@@ -67,7 +67,7 @@ public class CreateDagVisitorTest {
 
         a.localParallelism(1);
         dag.edge(between(a, b));
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(a.getLocalParallelism(), 1);
         assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
@@ -80,7 +80,7 @@ public class CreateDagVisitorTest {
         Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
 
         dag.edge(between(a, b));
-        CreateDagVisitor.decreaseParallelism(dag, 1);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 1);
 
         assertEquals(a.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
         assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
@@ -92,7 +92,7 @@ public class CreateDagVisitorTest {
         Vertex a = dag.newVertex("a", PROCESSOR_SUPPLIER);
         Vertex b = dag.newVertex("b", PROCESSOR_SUPPLIER);
         dag.edge(between(a, b));
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(Edge.RoutingPolicy.ISOLATED, dag.getOutboundEdges(a.getName()).get(0).getRoutingPolicy());
     }
@@ -106,7 +106,7 @@ public class CreateDagVisitorTest {
         Edge edge = dag.getOutboundEdges(a.getName()).get(0);
         edge.distributed();
 
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(a.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
         assertEquals(b.getLocalParallelism(), LOCAL_PARALLELISM_USE_DEFAULT);
@@ -122,7 +122,7 @@ public class CreateDagVisitorTest {
         edge.distributed();
         Edge.RoutingPolicy routingPolicy = edge.getRoutingPolicy();
 
-        CreateDagVisitor.decreaseParallelism(dag, 16);
+        CreateTopLevelDagVisitor.decreaseParallelism(dag, 16);
 
         assertEquals(routingPolicy, edge.getRoutingPolicy());
     }
